v0.0.6
- Possibly addressed issue #4
- The irc library used in this application didn't check if the file has
already been downloaded when a abort action occurd (which happens when
closing, since you want to abort dowloads which are still running when
you close the app). Now the library contains a check if the download is
still running before deleting a file.
- Possibly addressed issue #3
- Settings file was supposed to be put in the users/littleweeb folder,
but I guess some systems either don't allow to write there (admin
rights),  so now the settings file is located within the same directory
as the executable.

- Possibly fixed a unoticed issue where large downloads would sometimes
fail to download due to broken pipeline, possibly caused due to a
timeout check which halted the download to long, making the server you
connect to confused and closes the connection server side. This needs
more testing to be sure if it's fixed tho.

- Added a log function (it basically creates a new log file in the same
directory as the executable everytime you start the application). It
literally contains a copy of the console output, so in case of issues,
please add the log file!