v0.2.2
Small changes to how anime titles are parsed.

Issue reported here:
https://www.reddit.com/r/Piracy/comments/7ces5y/littleweeb_v02x_released_xdcc_done_easy_for_anime/dpr7lzt/

Cause: Creating synonyms for anime titles with just a single word would
create an empty search query, which retreived all the episodes available
on nibl, hence it looked like it crashed the interface while loading the
anime (but it was actually parsing thousands of episodes ^^).

Fix: Filtering the synonym array on duplicates and empty values.